Association of magic line ranges to system libraries is noe declarative.
Rather than being hard-coded in logic, there's now a table in perpet.c that
maps numeric-label ranges to system library file basenames. The sytem library
takes 1000-1999; the first new library to get this treatment is floatlib,
range 5000-5999.  The test for this is the ability to compile the pit's
fft.i program, which requires both syslib and floatlib.

Note that whenever you add a new library to this table, the build system
will need to be tweaked to add the library to the list of those copied to
the build directory.