Better fuzztest coverage, and move the fuzzer to etc/
The fuzztest now looks at idiotism.oil and tries to generate things
that look vaguely like the LHSes there, in addition to the normal
huge dose of randomness expected in a fuzztest. This makes it catch a
lot more bugs (i.e. it catches a greater proportion of bugs on each
run); it has found several optimizer bugs already, by the look of
things (although unfortunately fuzztests make it hard to track down
exactly where the bug is).