Handle missing lex and/or yacc correctly
This depends on the system in question having sane timestamps. Apparently,
DOS doesn't, and DOSBox even less so; I'm not sure what to do about this,
but this fix doesn't harm anything, and helps on systems like Linux and
OS X that are missing lex/flex and/or yacc/bison.