Specify background and foreground colors in ickstyle.css
Because before, it was specifying the background without specifying the
foreground; doing this either way round is a Cardinal Sin of Web
Accessibility, in case the user's default foreground happens to be set
to your specified background, etc.  The problem's solved here by
explicitly stating both background and foreground everywhere.